Fix force-update conditions
If the only reason not to update a fic is that the temp file has been updated more recently than the file in Calibre, then we should always import it, whether force is set or not.

If force == True and a fic has the same number of chapters online and in Calibre, we should update it. This also corresponds with what the help text for the force option already said.