Improve error handling and force-importing
Previously, we would export the work from calibre and try to update it with fanficfare. If appropriate, we would force-update it. Then we would add the work to calibre, set word count, and remove the previous work that was in calibre, even if we hadn't updated the file. This was unnecessary.

In the future, if we really want to update metadata of a previously-imported work, we can set force to true. (That means making two downloads, though - one for the original update attempt and one for the forced update. Hmm.)