Get list of series to reimport
Fics that are in more than one series seem to have a different series given in the front page of their epub than is listed in the Calibre metadata.
Get the ids of series that are affected this way, so we can reimport them (once the download script is updated to save all series fields from the AO3 data into Calibre).