chore: Remove the multiprocessing stuff
We were already not using multiprocessing (setting the number of processes in the pool to 1), because Calibre doesn't like being accessed by multiple processes at once. It could have been turned back on if not using Calibre, so I had kept the code here. Now we are adding in sleeps between downloading fics, if we get a Cloudflare error back, it makes even less sense to keep this complexity.