feat: Be less cautious about searching for newly-added book
I've never seen an error like the one we are trying to catch here, where we add a book and then get an error back when searching for it straight afterwards. Let's see if we can skip the try-except for it. (Maybe this was a problem when running multithreaded with Calibre?)