feat: Remove special handling of CloudflareWebsiteException
We currently work around this exception, which is caused by a misconfiguration between the AO3 site and Cloudflare, by using official AO3 mirror sites (archiveofourown.gay or archive.transformativeworks.org). These sites do not return the error response 525, so there is no need to catch and specially handle this exception in our code.